Major update to processor structure
This update includes an overhaul of the way the event processors
work within pixiescan. From now on, all processors will inherit
directly from EventProcessor (and not VandleProcessor). This will
make it easier to update processors in the future. Implemented
functional root output for Vandle and Liquid scintillators. This
update includes code from Mike Febbraro for the analysis of
liquid scintillator data.

This version of pixiescan outputs a root file named 'master.root'.
Each processor declares and handles its own local tree which
is written to this file upon exit. Each tree contains one
branch which, in turn, contains the variables of interest.
Root output added in this version...

 - Vandle (tof,lqdc,rqdc,tsLow,tsHigh,lMaxVal,rMaxVal,qdc,energy,
   multiplicity,location)

 - Liquid (TOF,S,L,liquid_tqdc,start_tqdc,location)
   S and L are the short and long pulse integrals respectively.