Changed the root/damm output system
Damm and Root output is now activated/deactivated using the boolean
flags 'use_root' and 'use_damm'. Added these flags to DetectorDriver,
EventProcessor, and TraceAnalyzer. Turning use_damm off will prevent
the program from declaring histograms in the .his file. This will
cut the size of the .his file and should save a little bit of time
and memory. For now, these flags are hard-coded within DetectorDriver.

Note: Turning all of the Damm histograms off causes the program to
quit unexpectedly due to zero bytes of histogram data being denied
for allocation by one of the libraries. I left one always turned on
to prevent this from happening ('channels with traces').

Changed the way the Root trees are filled. Responsibility of filling
processor trees has been moved to DetectorDriver (line 241). For each
event, DD will iterate through all processors. If the use_root flag is
set, DD will call the 'FillRoot()' method of each processor. This causes
all of the trees to have the same number of events and for those events
to be correlated. When the 'PackRoot' method is called, it sets a valid
flag to true. This valid flag is also stored in the root tree. If PackRoot
is not called before FillRoot, default values are saved to the tree, but
the valid flag is set to false.