Root data structure changes
I included ROOT dictionary generation in the Makefile. This should make
it easier to extract the data from the ROOT file in the future. It should
also make the filling of the tree more stable and less susceptible to
differences in computer architecture.

I also completely re-wrote the Makefile in order to make it slightly
more readable and to add the dictionary generation. At this time, you
need to add the path ./dict/ to the LD_LIBRARY_PATH in order for
the executable to find the shared library file.