Introduced saving of raw waveforms
Introduced a way to dump raw waveforms to the output root file instead
of the processed data. This is currently only implemented for the
LiquidProcessor, upon Mike Febbraro's request. The flag 'waveform'
may be passed from the command line when the executable is called, see
below. This flag will turn on the saving of the waveforms and turn off
the saving of the processed values.

Also introduced some flags which may be passed to the program when it is
called from the command line. Currently there are five recognized flags
which may be passed.

 - 'pfit-off': Turns pulse fitting off, it is ON by default.

 - 'dcfd-on': Turns the dcfd analyzer on, it is OFF by default.

 - 'root-off': Turns root file output off, it is ON by default.

 - 'damm-on': Turns damm output on, it is OFF by default.

 - 'waveform': Turns raw waveform output on, it is OFF by default.

These flags may be passed in any order but must come AFTER the first
argument (damm histogram name). Also changed the root output file to
use the histogram name instead of the generic name 'master'.

For example, './PixieLDF histname damm-on waveform' will fill the damm
histograms in file 'histname.his' and will dump raw waveform data to the
root file 'histname.root'.