Removed valid flag from data structures
Removed the valid flags from all classes in Structures.h. The multiplicity
variables may be used as a valid instead. If the multiplicity is zero, then
that processor had no valid events.

Added notes about turning Analyzers on.

Upon closing of root file, display the file size.

Added variable to DetectorDriver to track the total number of events. Now
track the total number of events and the number of events which are written
to the root tree.