Modified PulseAnalyzer/Viewer
Changed PulseAnalyzer and PulseViewer to take changes to the root
output format into account. Both tools will now expect a vector
of ints named 'something_wave' and an unsigned int with the name
'something_wave_mult'. The root name (something) is passed from
the command line. The program will automatically determine the
pulse size from these two branches.

Method 3 and 4 use the energy of the pulse for plotting. This energy
is obtained from a vector of doubles named 'something_energy'. Method
3 and 4 will not be usable if the program cannot find a branch with
this name.