Incorporated changes made by Zach Tully
Introduced new variables to the VandleProcessor. These variables
contain information about the recoil and ejectile particles
including x, y, and z position and angle and energy values.

These new variables required the modification of the VandleStructure
class and the modification of the timingCal and timingConstants
files.

I had to disable the saving of waveforms for Liquids and Triggers
because the compiler was complaining about the type that was being
passed (even though it worked fine before). I will fix it later.

I think that all of the new variables are working correctly. They
fill with values anyway. This code needs to be used with a correct
ldf file before I can know that it is working.