Added new IonChamberProcessor
Included a new IonChamberProcessor that accepts an "E" and a "dE"
preamp signal from an ion chamber. To include the ion chamber in
the map file, make the type "ion" and the subtype either "E" or
"dE" depending on the input signal.

This processor saves E, dE, and summed signals to the root file.