Major low-level data unpacking overhaul
Added a makefile flag 'USE_HHIRF' which may be used
to turn the use of HHIRF UPAK libraries on and off.
Setting this variable will build PixieScan as
before. Unsetting this variable eliminates the use
of scanor and damm and builds using libraries linked
from PixieSuite.

Includes a new main function which is intended to
replace the original scanor interface. This main
function is currently a placeholder but will be
modified in the future to include a full command
interface much like the original versions using
scanor.