Fixed fatal error which was causing seg-faults
In the previous commits, a function call from
FittingAnalyzer was being made after the memory for
the object was being freed. This error has been
corrected.

Added else statement in TimingInformation where the
user's timingCal.txt file is read in. Previously,
an entry in timingCal with all zeroes would cause
undefined behavior.

Changed the number of half words calculation back to
8 * nWords in pixieStd to alleviate data parsing
errors in hissub_sec.

Added r=1.0 for all entries in timingCal.txt.

Changed resOffset from 600 to 200 in
VandleProcessor. This value was causing some of the
damm histograms to not be filled.