Added SHM mode to PixieLDF
Added new command line options for PixieLDF including
SHM mode. SHM mode works just as scanning data from
a file, except that spills are obtained from the local
network instead of from a file.

ScanMain will wait for a spill to be dumped to the
local socket and will scan the spill after it
is received. Currently there is no way to stop the
program while it is listening to the SHM port.