Upgraded interface for new readout
Using the new readout scheme, files may now be scanned
by passing them from the command line. This eliminates
the need to run PixieScan and then type file and go.

i.e. ./PixieLDF output input.ldf

Will perform the same actions as:
./PixieLDF output
SCANOR-> file input.ldf
SCANOR-> go
SCANOR-> end

Shared-memory mode now brings up a new terminal
interface so that the "quit" commands may be
captured by the program. This allows for the
scanning of shm to be stopped at any time.

The USE_HHIRF flag DOES NOT work with the new
readout scheme at this time. There is currently
no way to pass the output filename prefix to the
damm initialization. The code does not crash, but
it aborts with the error message "SYNTAX ERROR IN
HIS-FILE SPECIFICATION". This hard abort also
causes the terminal to become corrupted. It is
currently not recommended to use the new readout
and damm together.