Added --dry-run option to PixieLDF
Added a new command line option (--dry-run) to
PixieLDF (new readout). This option will cause
the program to do everything normally except
send the data spills to the processors.

The new readout now keeps count of how many spills
are retrieved from the input file (or the network
in the case of shm) and will display this value
when the program exits.