Added --fast-fwd command to ScanMain
The --fast-fwd argument allows the user to skip a
specified number of words in the file. This is
useful for debugging a problematic portion of the
input file.

I found and fixed a potential memory leak in
ReadBuffData.RevD.cpp. I deleted the pointer to
the ChanEvent object which was defined using the
new operator but was never deleted.