Implemented checks for unusually large IDs
ScanList now checks that the pixie16 ID number is
within reasonable bounds before using the id to
access elements of a vector. This was causing
a crash whenever ScanList called vector::at with
an invalid id number.

ScanMain now ignores any spill which is marked as
invalid by DATA_buffer::Read. This was causing
crashes and headaches.

Removed the delete statement which was introduced
in the last commit. This was related to a memory
leak, but not the one I thought. It seems that
RemoveList was not being called for certain
conditions. I added several calls to RemoveList
from ReadSpill (I didn't change PixieStd.cpp
however). Upon return, the vector of pointers
to ChanEvents was being cleared, but the objects
they pointed to were never being deleted. This
should help with some of the memory leaks.