Added new .his file handler
Removed the old hhirf DAMM histogram file handler
and started implementing a new one.

Scan code interfaces with new handler in the same
ways as the old one i.e. hd1d_, hd2d_, and so on.

Code builds with no problems. Histogram file is
generated and extended with zeros, but calls
to count1cc_ are not implemented yet.