Histograms may now be added to output .his files
HisFile.cpp will now accept calls from the scan
code to add histograms to the output .his file.

These histograms may now also be filled with points,
but it is not clear if the ::flush() method is
filling the histograms correctly.

Added a new tool named hisReader. hisReader loads a
.drr file and displays the information of all
histograms stored inside it.