Fixed .his data storage error
Added new class HisData for storage of data read
in from .his files. Previously, data was being
accessed by using a short or int pointer to the
data character array but this does not always work
as intended.

Fixed the writing of histograms which are too short.
The size of histograms was being set to one bin too
small on both axes. This was causing histograms
displayed in damm to show data from following
histograms because it was reading in more bytes
than intended.