Modified the way new histos are appended to file
Pushing back the drr with a new histogram will now
extend the output .his file by writing a large
block of data, instead of writing individual
characters. This should prove to be faster for
large histograms.

Output .log file is now written upon closing the
histogram file, not when the .drr file is finalized.