Added compiler directives to ScanMain.cpp
Added several compiler directives to ScanMain in
order for it to be used in different ways. This
was done so that SimplePixieScan may use the
ScanMain source file from RootPixieScan instead
of relying on its own local version.

Also added directives which make it possible to
compile the source code without ncurses support.
This is useful for compiling on older machines
without the ncurses libraries. The normal scan
code will work, but SHM mode will NOT work without
ncurses at this time.