Changed RawEventStructure format
Changed RawEventStructure data format. Instead
of assigning a vector for each channel, the
module and channel numbers are recorded along
with the time and energy of the signal. This
was changed in order to move away from the
overly complicated code needed to view the
raw structure.

Brought the rawViewer tool into line with this
new format.

Added two command line arguments to ScanMain. They
are '--force-overwrite' and '--dump-raw-events'.
The former forces the scan code to overwrite the
output root file, if it exists. The latter dumps
raw event information to the file.