Added GenericStructure to root data structures
Added a new generic class to Structures.cpp. This
data structure is to be used for generic detector
types which don't fall into any other category.
This structure contains six vectors of doubles
(named generic_val1-6) for storing up to six
variables.

Fixed a crash related to attempting to write to
the output histogram file when it was not
initialized (i.e. USE_DAMM == false).

Uncommented writing of liquid waveforms to file. I
don't remember why it was inactive.

Added '--fast-fwd' command line argument to pulse
viewer. This makes it possible to skip immediately
to a pulse of interest.