ReadSpill now accepts an unsigned int array
Changed ReadSpill to accept an unsigned int
array instead of a character array. This was done
to prevent possible runtime errors by converting
a char array to an unsigned int array. This
conversion is NOT guaranteed to work in c++.

Added new command line argument for pulseViewer,
--mult-branch. This option may be used to specify
the name of the multiplicity branch to use. This
is useful for viewing waveforms with prefixes
which are different from their multiplicity (i.e.
left_wave and right_wave for VANDLE bars).